Corrigindo inconsistências de nomeclatura entre os scripts SQL e entities e também inconsistências de tipagem entre Long e Integer

+++ gisa/src/main/java/com/fatec/gisa/dtos/especialista/EspecialidadeDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class EspecialidadeDTO {
 
-    private Integer idEspecialidade;
+    private Long idEspecialidade;
     private String nome;
     private String descricao;
 }
